refactor: simplify portfolio funding component and fiscal year handling
