+++ frontend/src/components/Portfolios/PortfolioFunding/index.js
@@ -1,0 +1,1 @@
+export { default } from "./PortfolioFunding";
